chore(site): rewrite style imports using root, adds search param pagination
