+++ app/collection/components/collection-filter/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .collectionFilter {
   margin-block-end: 3rem;
