+++ app/collection/components/collection-item-metadata/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .collectionItemData {
   padding: 0.25rem 0.5rem;
