+++ app/collection/components/collection-item/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .collectionItem {
   background-color: rgba(vars.$tertiary-color-1, 0.15);
