+++ app/collection/components/collection-item/types.d.ts
@@ -1,4 +1,4 @@
-import { Game } from "types/games"
+import { Game } from "app/types/games"
 
 export type CollectionItemProps = {
   data: Game
