+++ app/collection/components/collection-list/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .collectionList {
   display: grid;
