+++ app/components/button/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .button {
   height: 100%;
