+++ app/components/pagination/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .pagination {
   display: flex;
