+++ app/components/site-footer/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .pageFooter {
   border: 2px solid vars.$tertiary-color-2;
