+++ app/components/site-header/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .logoContainer {
   display: inline-block;
