+++ app/essays/[slug]/not-found.tsx
@@ -1,4 +1,4 @@
-import { Layout } from "components/layout"
+import { Layout } from "app/components/layout"
 
 export default function NotFound() {
   return (
