+++ app/essays/components/essay-content/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .essayContent {
   line-height: 2;
