+++ app/essays/components/essay-footer/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .essayFooter {
   color: rgba(vars.$text-color, 0.7);
