+++ app/essays/components/essay-list-item/types.d.ts
@@ -1,4 +1,4 @@
-import { Essay } from "types/essays"
+import { Essay } from "app/types/essays"
 
 export type EssayListItemProps = {
   entry: Essay
