+++ app/essays/components/essay-list-loader/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/vars" as vars;
+@use "app/styles/vars" as vars;
 
 .empyStateContainer {
   margin-block-end: 3rem;
