+++ app/essays/components/essay-list/types.d.ts
@@ -1,4 +1,4 @@
-import { EssayPageData } from "types/essays"
+import { EssayPageData } from "app/types/essays"
 
 export type EssayListComponentProps = {
   data: EssayPageData
