+++ app/essays/page.tsx
@@ -1,8 +1,8 @@
 import { PageHeading, Layout } from "app/components"
 import { Essays } from "./components/essays"
-import { EssayPageData } from "types/essays"
-import { getEssayList } from "lib/get-essay-list"
-import { BASE_TITLE } from "lib/constants"
+import { EssayPageData } from "app/types/essays"
+import { getEssayList } from "app/utils/get-essay-list"
+import { BASE_TITLE } from "app/utils/constants"
 
 const DEFAULT_PAGE = 0
 
@@ -14,8 +14,15 @@ export const metadata = {
   description: "Essays about video games, collecting, and nonsense",
 }
 
-export default async function Page() {
-  const initialData = (await getEssayList(DEFAULT_PAGE)) as EssayPageData
+export default async function Page({
+  searchParams: { page },
+}: {
+  searchParams: { page: string }
+}) {
+  const pageNum = typeof page === "string" ? parseInt(page) : undefined
+  const resolvedPageNum =
+    typeof pageNum === "number" ? pageNum - 1 : DEFAULT_PAGE
+  const initialData = (await getEssayList(resolvedPageNum)) as EssayPageData
 
   return (
     <Layout>
