+++ app/essays/utils/types.d.ts
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from "react"
-import { EssayPageData } from "types/essays"
+import { EssayPageData } from "app/types/essays"
 
 export interface FetchEssaysArgs {
   index: number
