+++ tsconfig.json
@@ -16,13 +16,7 @@
     "incremental": true,
     "baseUrl": ".",
     "paths": {
-      // Public and lib must be in the root
-      "public/*": ["public/*"],
-      // App stuff
-      "lib/*": ["lib/*"],
-      "styles/*": ["styles/*"],
-      "components/*": ["app/components/*"],
-      "types/*": ["types/*"]
+      "public/*": ["public/*"]
     },
     "plugins": [
       {
